requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
cryptography>=42.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
jinja2>=3.1.0
